+++ .gitignore
@@ -1,3 +1,6 @@
+# next
+.next
+
 # dependencies
 /node_modules
