--- Caddyfile
@@ -1,17 +1,0 @@
-# Global options (See https://caddyserver.com/docs/caddyfile/options)
-{
-  # Disable the admin API
-  admin off
-}
-
-# Site options
-:80 {
-  # Single page application redirect
-  try_files {path} /index.html
-
-  # Filesystem location
-  root * {$STATIC}
-
-  # Static file server
-  file_server
-}
