+++ package.json
@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
   },
   "dependencies": {
     "@apollo/client": "^3.3.19",
@@ -17,7 +16,6 @@
     "@types/node": "^14.14.41",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
     "@types/react-simple-maps": "^1.0.3",
     "@types/react-syntax-highlighter": "^13.5.0",
     "chart.js": "^3.2.1",
@@ -27,6 +25,10 @@
     "fuse.js": "^6.4.6",
     "graphql": "^15.5.0",
     "luxon": "^1.26.0",
+    "next": "^10.2.3",
+    "next-compose-plugins": "^2.2.1",
+    "next-images": "^1.7.0",
+    "next-transpile-modules": "^7.2.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-chartjs-2": "^3.0.3",
@@ -34,8 +36,6 @@
     "react-feather": "^2.0.9",
     "react-markdown": "^6.0.0",
     "react-particles-js": "^3.4.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
     "react-simple-maps": "^2.3.0",
     "react-spring": "^9.1.2",
     "react-syntax-highlighter": "^15.4.3",
