--- src/App.module.css
@@ -1,9 +1,0 @@
-/* Particles */
-.Particles {
-  height: 100%;
-  left: 0;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: -4;
-}
