+++ src/components/Error/Error.test.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import About from './About';
+import Error from './Error';
 
-describe('<About />', () => {
+describe('<Error />', () => {
   test('it should mount', () => {
-    render(<About />);
+    render(<Error />);
     
-    const about = screen.getByTestId('About');
+    const error = screen.getByTestId('Error');
 
-    expect(about).toBeInTheDocument();
+    expect(error).toBeInTheDocument();
   });
 });
