--- src/pages/Scoreboard/Scoreboard.module.css
@@ -1,10 +1,0 @@
-/* Graph */
-.graph {
-  width: 60vw;
-}
-
-/* Table row */
-.row > * {
-  border-color: var(--chakra-colors-gray-500) !important;
-  padding: 10px !important;
-}
